Move zuora common utilities to utils/common directory
Moved zuora/src/common.ts to modules/zuora/src/utils/common.ts and updated all imports to reference the new path. This improves code organization by placing shared utilities in a dedicated utils directory.
