+++ modules/zuora/src/billingPreview.ts
@@ -2,7 +2,7 @@ import { groupBy, sortBy, sumNumbers } from '@modules/arrayFunctions';
 import { getIfDefined } from '@modules/nullAndUndefined';
 import dayjs from 'dayjs';
 import type { Dayjs } from 'dayjs';
-import { zuoraDateFormat } from './common';
+import { zuoraDateFormat } from './utils/common';
 import type { ZuoraClient } from './zuoraClient';
 import type { BillingPreview } from './zuoraSchemas';
 import { billingPreviewSchema } from './zuoraSchemas';
