+++ modules/zuora/src/payment.ts
@@ -1,5 +1,5 @@
 import type { Dayjs } from 'dayjs';
-import { zuoraDateFormat } from '@modules/zuora/common';
+import { zuoraDateFormat } from '@modules/zuora/utils/common';
 import type { ZuoraClient } from '@modules/zuora/zuoraClient';
 import type { ZuoraUpperCaseSuccessResponse } from '@modules/zuora/zuoraSchemas';
 import { zuoraUpperCaseSuccessResponseSchema } from '@modules/zuora/zuoraSchemas';
