+++ modules/zuora/src/zuoraClient.ts
@@ -3,7 +3,7 @@ import type { Stage } from '@modules/stage';
 import { Logger } from '@modules/logger';
 import { BearerTokenProvider } from './auth/bearerTokenProvider';
 import { getOAuthClientCredentials } from './auth/oAuthCredentials';
-import { zuoraServerUrl } from './common';
+import { zuoraServerUrl } from './utils/common';
 import { generateZuoraError } from './errors/zuoraErrorHandler';
 import type { ZuoraResponse } from './types/httpResponse';
