+++ modules/zuora/test/fixtures/request-bodies/supporterplus-subscribe-body-tier2.ts
@@ -1,5 +1,5 @@
 import type { Dayjs } from 'dayjs';
-import { zuoraDateFormat } from '@modules/zuora/common';
+import { zuoraDateFormat } from '@modules/zuora/utils/common';
 import { catalog } from '../../../../../handlers/discount-api/src/productToDiscountMapping';
 
 export const supporterPlusSubscribeBody = (subscriptionDate: Dayjs) => {
